Successful Build but Running causes Exceptions
Exceptions due new Security changes of iPhone Apps:
'Application windows are expected to have a root view controller at the
end of application launch'

….fixing in progress:

+++ xPTAMclean/iCVD/thread.cpp
@@ -76,7 +76,7 @@ Thread* Thread::getCurrent()
 }
 
 /** Uses nanosleep. */
-void Thread::sleep(unsigned int milli) 
+void Thread::sleep(long milli)
 {
    struct timespec ts = { milli/1000, (milli%1000)*1000000 };
    nanosleep(&ts, 0);
